updating to new version of libcolor and removing my types.h

+++ include/loguva.h
@@ -4,8 +4,6 @@
 #include <stdbool.h>
 #include <stdio.h>
 
-#include "types.h"
-
 #define loguva(lv, fmt, ...)                                                   \
   loguva_log(lv, __FILE__, __LINE__, fmt, ##__VA_ARGS__)
 
@@ -13,7 +11,8 @@ enum log_levels { DEBUG, INFO, WARNING, ERROR, FATAL };
 
 bool loguva_add_stream(FILE *stream);
 FILE *loguva_add_file(const char *path);
-bool loguva_log(enum log_levels lv, const char *path, u64 line, const char *fmt,
-                ...) __attribute__((format(printf, 4, 5)));
+
+bool loguva_log(enum log_levels lv, const char *path, size_t line,
+                const char *fmt, ...) __attribute__((format(printf, 4, 5)));
 
 #endif
